+++ apps/cli/cli_common.c
@@ -695,13 +695,13 @@ compare_dbs(clicon_handle h,
 	astext = cv_int32_get(cvec_i(argv, 0));
     else
 	astext = 0;
-    if (clicon_rpc_get_config(h, "running", "/", NULL, &xc1) < 0)
+    if (clicon_rpc_get_config(h, NULL, "running", "/", NULL, &xc1) < 0)
 	goto done;
     if ((xerr = xpath_first(xc1, "/rpc-error")) != NULL){
 	clicon_rpc_generate_error("Get configuration", xerr);
 	goto done;
     }
-    if (clicon_rpc_get_config(h, "candidate", "/", NULL, &xc2) < 0)
+    if (clicon_rpc_get_config(h, NULL, "candidate", "/", NULL, &xc2) < 0)
 	goto done;
     if ((xerr = xpath_first(xc2, "/rpc-error")) != NULL){
 	clicon_rpc_generate_error("Get configuration", xerr);
@@ -862,7 +862,7 @@ save_config_file(clicon_handle h,
 	goto done;
     }
     filename = cv_string_get(cv);
-    if (clicon_rpc_get_config(h, dbstr,"/", NULL, &xt) < 0)
+    if (clicon_rpc_get_config(h, NULL, dbstr,"/", NULL, &xt) < 0)
 	goto done;
     if (xt == NULL){
 	clicon_err(OE_CFG, 0, "get config: empty tree"); /* Shouldnt happen */
@@ -1203,8 +1203,10 @@ cli_copy_config(clicon_handle h,
 	goto done;
     }
     cprintf(cb, xpath, keyname, fromname);	
+    if ((nsc = xml_nsctx_init(NULL, namespace)) == NULL)
+	goto done;
     /* Get from object configuration and store in x1 */
-    if (clicon_rpc_get_config(h, db, cbuf_get(cb), namespace, &x1) < 0)
+    if (clicon_rpc_get_config(h, NULL, db, cbuf_get(cb), nsc, &x1) < 0)
 	goto done;
     if ((xerr = xpath_first(x1, "/rpc-error")) != NULL){
 	clicon_rpc_generate_error("Get configuration", xerr);
@@ -1224,8 +1226,7 @@ cli_copy_config(clicon_handle h,
 	goto done;
     xml_name_set(x2, "config");
     cprintf(cb, "/%s", keyname);	
-    if ((nsc = xml_nsctx_init(NULL, namespace)) == NULL)
-	goto done;
+
     if ((x = xpath_first_nsc(x2, nsc, "%s", cbuf_get(cb))) == NULL){
 	clicon_err(OE_PLUGIN, 0, "Field %s not found in copy tree", keyname);
 	goto done;